+++ Cargo.toml
@@ -16,7 +16,6 @@ members = [
     "config",
     "bus",
     "storage",
-    "network",
     "consensus",
     "txpool",
     "rpc/json-rpc",
@@ -26,10 +25,8 @@ members = [
     "sync",
     "miner",
     "node",
-    "network-libp2p",
-    "network-libp2p/peerset",
     "network-p2p",
     "network-p2p/peerset",
-    "network-new",
+    "network",
     "vm/vm-runtime",
 ]
