--- network-libp2p/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "network-libp2p"
-version = "0.1.0"
-authors = ["Starcoin Core Dev <dev@starcoin.org>"]
-license = "Apache-2.0"
-publish = false
-edition = "2018"
-
-[dependencies]
-byteorder = "1.3"
-bytes = "0.4"
-fnv = "1.0"
-futures = "0.1"
-libp2p = { git ="https://github.com/starcoinorg/rust-libp2p", branch = "starcoin"}
-parking_lot = "0.7.1"
-lazy_static = "1.2"
-log = "0.4"
-rand = "0.6"
-serde = { version = "1.0.70", features = ["derive"] }
-serde_json = "1.0.24"
-smallvec = "0.6"
-sc-peerset = { version = "2.0.0", path = "peerset" }
-tokio = "0.1"
-tokio-io = "0.1"
-tokio-timer = "0.2"
-unsigned-varint = { version = "0.2.1", features = ["codec"] }
-void = "1.0"
-futures03 = { package = "futures", version = "0.3.1", features = ["compat"] }
-futures-timer = "0.4.0"
-zeroize = "1.0.0"
-
-[dev-dependencies]
-tempfile = "3.1.0"
