--- network-libp2p/peerset/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-description = "Connectivity manager based on reputation"
-homepage = "http://parity.io"
-license = "GPL-3.0"
-name = "sc-peerset"
-version = "2.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-
-[dependencies]
-futures = "0.3.1"
-libp2p = { git ="https://github.com/starcoinorg/rust-libp2p",branch = "starcoin", default-features = false}
-
-log = "0.4.8"
-serde_json = "1.0.41"
-
-[dev-dependencies]
-rand = "0.7.2"
