--- network-libp2p/src/custom_proto.rs
@@ -1,8 +1,0 @@
-// Copyright (c) The Starcoin Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub use behaviour::{CustomProto, CustomProtoOut};
-mod behaviour;
-mod handler;
-mod tests;
-mod upgrade;
