--- network-libp2p/src/custom_proto/upgrade.rs
@@ -1,276 +1,0 @@
-// Copyright (c) The Starcoin Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-use crate::ProtocolId;
-use bytes::{Bytes, BytesMut};
-use futures::{future, prelude::*, stream};
-use libp2p::core::{
-    upgrade::ProtocolName, Endpoint, InboundUpgrade, Negotiated, OutboundUpgrade, UpgradeInfo,
-};
-use libp2p::tokio_codec::Framed;
-use std::{collections::VecDeque, io, vec::IntoIter as VecIntoIter};
-use tokio_io::{AsyncRead, AsyncWrite};
-use unsigned_varint::codec::UviBytes;
-
-/// Connection upgrade for a single protocol.
-///
-/// Note that "a single protocol" here refers to `par` for example. However
-/// each protocol can have multiple different versions for networking purposes.
-pub struct RegisteredProtocol {
-    /// Id of the protocol for API purposes.
-    id: ProtocolId,
-    /// Base name of the protocol as advertised on the network.
-    /// Ends with `/` so that we can append a version number behind.
-    base_name: Bytes,
-    /// List of protocol versions that we support.
-    /// Ordered in descending order so that the best comes first.
-    supported_versions: Vec<u8>,
-}
-
-impl RegisteredProtocol {
-    /// Creates a new `RegisteredProtocol`. The `custom_data` parameter will be
-    /// passed inside the `RegisteredProtocolOutput`.
-    pub fn new(protocol: impl Into<ProtocolId>, versions: &[u8]) -> Self {
-        let protocol = protocol.into();
-        let mut base_name = Bytes::from_static(b"/stargate/");
-        base_name.extend_from_slice(protocol.as_bytes());
-        base_name.extend_from_slice(b"/");
-
-        RegisteredProtocol {
-            base_name,
-            id: protocol,
-            supported_versions: {
-                let mut tmp = versions.to_vec();
-                tmp.sort_unstable_by(|a, b| b.cmp(&a));
-                tmp
-            },
-        }
-    }
-}
-
-impl Clone for RegisteredProtocol {
-    fn clone(&self) -> Self {
-        RegisteredProtocol {
-            id: self.id.clone(),
-            base_name: self.base_name.clone(),
-            supported_versions: self.supported_versions.clone(),
-        }
-    }
-}
-
-/// Output of a `RegisteredProtocol` upgrade.
-pub struct RegisteredProtocolSubstream<TSubstream> {
-    /// If true, we are in the process of closing the sink.
-    is_closing: bool,
-    /// Whether the local node opened this substream (dialer), or we received this substream from
-    /// the remote (listener).
-    endpoint: Endpoint,
-    /// Buffer of packets to send.
-    send_queue: VecDeque<Vec<u8>>,
-    /// If true, we should call `poll_complete` on the inner sink.
-    requires_poll_complete: bool,
-    /// The underlying substream.
-    inner: stream::Fuse<Framed<Negotiated<TSubstream>, UviBytes<Vec<u8>>>>,
-    /// Version of the protocol that was negotiated.
-    protocol_version: u8,
-    /// If true, we have sent a "remote is clogged" event recently and shouldn't send another one
-    /// unless the buffer empties then fills itself again.
-    clogged_fuse: bool,
-}
-
-impl<TSubstream> RegisteredProtocolSubstream<TSubstream> {
-    /// Returns the version of the protocol that was negotiated.
-    pub fn protocol_version(&self) -> u8 {
-        self.protocol_version
-    }
-
-    /// Returns whether the local node opened this substream (dialer), or we received this
-    /// substream from the remote (listener).
-    pub fn endpoint(&self) -> Endpoint {
-        self.endpoint
-    }
-
-    /// Starts a graceful shutdown process on this substream.
-    ///
-    /// Note that "graceful" means that we sent a closing message. We don't wait for any
-    /// confirmation from the remote.
-    ///
-    /// After calling this, the stream is guaranteed to finish soon-ish.
-    pub fn shutdown(&mut self) {
-        self.is_closing = true;
-        self.send_queue.clear();
-    }
-
-    /// Sends a message to the substream.
-    pub fn send_message(&mut self, data: Vec<u8>) {
-        if self.is_closing {
-            return;
-        }
-
-        self.send_queue.push_back(data);
-    }
-}
-
-/// Event produced by the `RegisteredProtocolSubstream`.
-#[derive(Debug, Clone)]
-pub enum RegisteredProtocolEvent {
-    /// Received a message from the remote.
-    Message(BytesMut),
-
-    /// Diagnostic event indicating that the connection is clogged and we should avoid sending too
-    /// many messages to it.
-    Clogged {
-        /// Copy of the messages that are within the buffer, for further diagnostic.
-        messages: Vec<Vec<u8>>,
-    },
-}
-
-impl<TSubstream> Stream for RegisteredProtocolSubstream<TSubstream>
-where
-    TSubstream: AsyncRead + AsyncWrite,
-{
-    type Item = RegisteredProtocolEvent;
-    type Error = io::Error;
-
-    fn poll(&mut self) -> Poll<Option<Self::Item>, Self::Error> {
-        // Flushing the local queue.
-        while let Some(packet) = self.send_queue.pop_front() {
-            match self.inner.start_send(packet)? {
-                AsyncSink::NotReady(packet) => {
-                    self.send_queue.push_front(packet);
-                    break;
-                }
-                AsyncSink::Ready => self.requires_poll_complete = true,
-            }
-        }
-
-        // If we are closing, close as soon as the Sink is closed.
-        if self.is_closing {
-            return Ok(self.inner.close()?.map(|()| None));
-        }
-
-        // Indicating that the remote is clogged if that's the case.
-        if self.send_queue.len() >= 2048 {
-            if !self.clogged_fuse {
-                // Note: this fuse is important not just for preventing us from flooding the logs;
-                // 	if you remove the fuse, then we will always return early from this function and
-                //	thus never read any message from the network.
-                self.clogged_fuse = true;
-                return Ok(Async::Ready(Some(RegisteredProtocolEvent::Clogged {
-                    messages: self.send_queue.iter().map(|m| m.clone()).collect(),
-                })));
-            }
-        } else {
-            self.clogged_fuse = false;
-        }
-
-        // Flushing if necessary.
-        if self.requires_poll_complete {
-            if let Async::Ready(()) = self.inner.poll_complete()? {
-                self.requires_poll_complete = false;
-            }
-        }
-
-        // Receiving incoming packets.
-        // Note that `inner` is wrapped in a `Fuse`, therefore we can poll it forever.
-        match self.inner.poll()? {
-            Async::Ready(Some(data)) => {
-                Ok(Async::Ready(Some(RegisteredProtocolEvent::Message(data))))
-            }
-            Async::Ready(None) => {
-                if !self.requires_poll_complete && self.send_queue.is_empty() {
-                    Ok(Async::Ready(None))
-                } else {
-                    Ok(Async::NotReady)
-                }
-            }
-            Async::NotReady => Ok(Async::NotReady),
-        }
-    }
-}
-
-impl UpgradeInfo for RegisteredProtocol {
-    type Info = RegisteredProtocolName;
-    type InfoIter = VecIntoIter<Self::Info>;
-
-    #[inline]
-    fn protocol_info(&self) -> Self::InfoIter {
-        // Report each version as an individual protocol.
-        self.supported_versions
-            .iter()
-            .map(|&version| {
-                let num = version.to_string();
-
-                let mut name = self.base_name.clone();
-                name.extend_from_slice(num.as_bytes());
-                RegisteredProtocolName { name, version }
-            })
-            .collect::<Vec<_>>()
-            .into_iter()
-    }
-}
-
-/// Implementation of `ProtocolName` for a custom protocol.
-#[derive(Debug, Clone)]
-pub struct RegisteredProtocolName {
-    /// Protocol name, as advertised on the wire.
-    name: Bytes,
-    /// Version number. Stored in string form in `name`, but duplicated here for easier retrieval.
-    version: u8,
-}
-
-impl ProtocolName for RegisteredProtocolName {
-    fn protocol_name(&self) -> &[u8] {
-        &self.name
-    }
-}
-
-impl<TSubstream> InboundUpgrade<TSubstream> for RegisteredProtocol
-where
-    TSubstream: AsyncRead + AsyncWrite,
-{
-    type Output = RegisteredProtocolSubstream<TSubstream>;
-    type Future = future::FutureResult<Self::Output, io::Error>;
-    type Error = io::Error;
-
-    fn upgrade_inbound(self, socket: Negotiated<TSubstream>, info: Self::Info) -> Self::Future {
-        let framed = {
-            let mut codec = UviBytes::default();
-            codec.set_max_len(16 * 1024 * 1024); // 16 MiB hard limit for packets.
-            Framed::new(socket, codec)
-        };
-
-        future::ok(RegisteredProtocolSubstream {
-            is_closing: false,
-            endpoint: Endpoint::Listener,
-            send_queue: VecDeque::new(),
-            requires_poll_complete: false,
-            inner: framed.fuse(),
-            protocol_version: info.version,
-            clogged_fuse: false,
-        })
-    }
-}
-
-impl<TSubstream> OutboundUpgrade<TSubstream> for RegisteredProtocol
-where
-    TSubstream: AsyncRead + AsyncWrite,
-{
-    type Output = <Self as InboundUpgrade<TSubstream>>::Output;
-    type Future = <Self as InboundUpgrade<TSubstream>>::Future;
-    type Error = <Self as InboundUpgrade<TSubstream>>::Error;
-
-    fn upgrade_outbound(self, socket: Negotiated<TSubstream>, info: Self::Info) -> Self::Future {
-        let framed = Framed::new(socket, UviBytes::default());
-
-        future::ok(RegisteredProtocolSubstream {
-            is_closing: false,
-            endpoint: Endpoint::Dialer,
-            send_queue: VecDeque::new(),
-            requires_poll_complete: false,
-            inner: framed.fuse(),
-            protocol_version: info.version,
-            clogged_fuse: false,
-        })
-    }
-}
