--- network-new/src/lib.rs
@@ -1,23 +1,0 @@
-// Copyright (c) The Starcoin Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-#[macro_use]
-extern crate log;
-
-mod helper;
-mod message_processor;
-mod messages;
-mod net;
-mod net_test;
-mod network;
-mod sync_messages;
-
-pub use messages::*;
-
-pub use helper::{
-    convert_account_address_to_peer_id, convert_peer_id_to_account_address, get_unix_ts,
-};
-
-use anyhow::{Error, Result};
-pub use network_p2p::PeerId;
-use types::system_events::SystemEvents;
