+++ network/src/net.rs
@@ -11,43 +11,264 @@ use crypto::{
 };
 
 use crate::messages::{Message, NetworkMessage};
+use anyhow::*;
 use futures::{
-    future,
-    stream::{self, Stream},
-    sync::{
+    channel::{
         mpsc,
         oneshot::{self, Canceled, Sender},
     },
-    try_ready, Async, Future,
+    prelude::*,
 };
 
+use anyhow::*;
+use bytes::Bytes;
 use config::NetworkConfig;
-use network_libp2p::{
-    identity, start_service, NetworkConfiguration, NodeKeyConfig, Secret, Service as Libp2pService,
-    ServiceEvent,
+use libp2p::PeerId;
+use network_p2p::{
+    identity, Event, GenericProtoOut as ServiceEvent, NetworkConfiguration, NetworkService,
+    NetworkWorker, NodeKeyConfig, Params, Secret,
 };
 use parity_codec::alloc::collections::HashSet;
 use parking_lot::Mutex;
 use scs::SCSCodec;
+use slog::Drain;
+use std::cell::{Cell, RefCell};
+use std::task::{Context, Poll};
 use std::{collections::HashMap, io, sync::Arc, thread};
-use tokio::prelude::task::AtomicTask;
+use tokio::runtime::Handle;
 use types::account_address::AccountAddress;
 
 #[derive(Clone)]
-pub struct NetworkService {
-    pub libp2p_service: Arc<Mutex<Libp2pService>>,
+pub struct SNetworkService {
+    handle: Handle,
+    inner: NetworkInner,
+    service: Arc<NetworkService>,
+    net_tx: Option<mpsc::UnboundedSender<NetworkMessage>>,
+}
+
+#[derive(Clone)]
+pub struct NetworkInner {
+    service: Arc<NetworkService>,
     acks: Arc<Mutex<HashMap<u128, Sender<()>>>>,
+}
+
+impl SNetworkService {
+    pub fn new(cfg: NetworkConfiguration, handle: Handle) -> Self {
+        let protocol = network_p2p::ProtocolId::from("stargate".as_bytes());
+
+        let worker = NetworkWorker::new(Params::new(cfg, protocol)).unwrap();
+        let service = worker.service().clone();
+        let worker = worker;
+
+        let acks = Arc::new(Mutex::new(HashMap::new()));
+
+        handle.spawn(worker);
+
+        let inner = NetworkInner {
+            service: service.clone(),
+            acks,
+        };
+
+        Self {
+            inner,
+            handle,
+            service,
+            net_tx: None,
+        }
+    }
+
+    pub fn run(
+        &mut self,
+    ) -> (
+        mpsc::UnboundedSender<NetworkMessage>,
+        mpsc::UnboundedReceiver<NetworkMessage>,
+        mpsc::UnboundedReceiver<PeerEvent>,
+        mpsc::UnboundedSender<()>,
+    ) {
+        let (close_tx, close_rx) = mpsc::unbounded::<()>();
+        let (mut tx, net_rx) = mpsc::unbounded();
+        let (net_tx, mut rx) = mpsc::unbounded::<NetworkMessage>();
+        let (event_tx, mut event_rx) = mpsc::unbounded::<PeerEvent>();
+        let inner = self.inner.clone();
+
+        self.net_tx = Some(net_tx.clone());
+
+        self.handle
+            .spawn(Self::start_network(inner, tx, rx, event_tx, close_rx));
+        (net_tx, net_rx, event_rx, close_tx)
+    }
+
+    async fn start_network(
+        inner: NetworkInner,
+        net_tx: mpsc::UnboundedSender<NetworkMessage>,
+        mut net_rx: mpsc::UnboundedReceiver<NetworkMessage>,
+        event_tx: mpsc::UnboundedSender<PeerEvent>,
+        mut close_rx: mpsc::UnboundedReceiver<()>,
+    ) {
+        let mut event_stream = inner.service.event_stream().fuse();
+        let mut net_rx = net_rx.fuse();
+        let mut close_rx = close_rx.fuse();
+
+        loop {
+            futures::select! {
+                message = net_rx.select_next_some()=>{
+                    inner.handle_network_send(message).await.unwrap();
+                    info!("send net message");
+                },
+                event = event_stream.select_next_some()=>{
+                    inner.handle_network_receive(event,net_tx.clone(),event_tx.clone()).await.unwrap();
+                    info!("receive net message");
+                },
+                _ = close_rx.select_next_some() => {
+                    debug!("To shutdown command ");
+                    break;
+                }
+                complete => {
+                    warn!("all stream are complete");
+                    break;
+                }
+            }
+        }
+    }
+
+    pub async fn is_connected(&self, address: AccountAddress) -> Result<bool> {
+        let peer_id = convert_account_address_to_peer_id(address)?;
+        Ok(self.service.is_connected(peer_id).await)
+    }
+
+    pub fn identify(&self) -> AccountAddress {
+        convert_peer_id_to_account_address(self.service.peer_id()).unwrap()
+    }
+
+    pub async fn send_message(
+        &self,
+        account_address: AccountAddress,
+        message: Vec<u8>,
+    ) -> Result<()> {
+        let (tx, rx) = oneshot::channel::<()>();
+        let (protocol_msg, message_id) = Message::new_payload(message);
+        let peer_id =
+            convert_account_address_to_peer_id(account_address).expect("Invalid account address");
+
+        info!("send message is  {:?}", protocol_msg);
+        self.service
+            .send_notification(peer_id, protocol_msg.into_bytes());
+        info!("Send message with ack");
+        //self.waker.wake();
+        self.inner.acks.lock().insert(message_id, tx);
+        rx.await?;
+
+        Ok(())
+    }
+
+    pub async fn broadcast_message(&mut self, message: Vec<u8>) {
+        info!("broadcast message");
+        let (protocol_msg, message_id) = Message::new_payload(message);
+
+        let message_bytes = protocol_msg.into_bytes();
+
+        self.service.broadcast_message(message_bytes).await;
+    }
+
+    pub async fn connected_peers(&self) -> HashSet<PeerId> {
+        self.service.connected_peers().await
+    }
+}
+
+impl NetworkInner {
+    async fn handle_network_receive(
+        &self,
+        event: Event,
+        net_tx: mpsc::UnboundedSender<NetworkMessage>,
+        event_tx: mpsc::UnboundedSender<PeerEvent>,
+    ) -> Result<()> {
+        info!("message is {:?}", event);
+        match event {
+            Event::Dht(_) => {
+                info!("ignore dht event");
+            }
+            Event::NotificationStreamOpened { remote } => {
+                let addr = convert_peer_id_to_account_address(&remote)?;
+                info!("Connected peer {:?}", addr);
+                let open_msg = PeerEvent::Open(addr);
+                event_tx.unbounded_send(open_msg)?;
+            }
+            Event::NotificationStreamClosed { remote } => {
+                let addr = convert_peer_id_to_account_address(&remote)?;
+                info!("Close peer {:?}", addr);
+                let open_msg = PeerEvent::Close(addr);
+                event_tx.unbounded_send(open_msg)?;
+            }
+            Event::NotificationsReceived { remote, messages } => {
+                self.handle_messages(remote, messages, net_tx).await?;
+            }
+        }
+        Ok(())
+    }
+
+    async fn handle_messages(
+        &self,
+        peer_id: PeerId,
+        messages: Vec<Bytes>,
+        net_tx: mpsc::UnboundedSender<NetworkMessage>,
+    ) -> Result<()> {
+        info!("Receive message with peer_id:{:?}", &peer_id);
+        let address = convert_peer_id_to_account_address(&peer_id)?;
+        info!("addr is {:?}", address);
+        for message in messages {
+            let message = Message::from_bytes(message.as_ref())?;
+            info!("message is {:?}", message);
+            match message {
+                Message::Payload(payload) => {
+                    //receive message
+                    let user_msg = NetworkMessage {
+                        peer_id: address,
+                        data: payload.data,
+                    };
+                    net_tx.unbounded_send(user_msg)?;
+                    if payload.id != 0 {
+                        self.service.send_notification(
+                            peer_id.clone(),
+                            Message::ACK(payload.id).into_bytes(),
+                        );
+                    }
+                }
+                Message::ACK(message_id) => {
+                    info!("Receive message ack");
+                    if let Some(mut tx) = self.acks.lock().remove(&message_id) {
+                        let _ = tx.send(());
+                    } else {
+                        error!(
+                            "Receive a invalid ack, message id:{}, peer id:{}",
+                            message_id, peer_id
+                        );
+                    }
+                }
+            }
+        }
+        Ok(())
+    }
+
+    async fn handle_network_send(&self, message: NetworkMessage) -> Result<()> {
+        let account_addr = message.peer_id.clone();
+        self.service.send_notification(
+            convert_account_address_to_peer_id(account_addr)?,
+            Message::new_payload(message.data).0.into_bytes(),
+        );
+        Ok(())
+    }
 }
 
 pub fn build_network_service(
     cfg: &NetworkConfig,
     key_pair: Arc<KeyPair<Ed25519PrivateKey, Ed25519PublicKey>>,
+    handle: Handle,
 ) -> (
-    NetworkService,
+    SNetworkService,
     mpsc::UnboundedSender<NetworkMessage>,
     mpsc::UnboundedReceiver<NetworkMessage>,
     mpsc::UnboundedReceiver<PeerEvent>,
-    oneshot::Sender<()>,
+    mpsc::UnboundedSender<()>,
 ) {
     let config = NetworkConfiguration {
         listen_addresses: vec![cfg.listen.parse().expect("Failed to parse network config")],
@@ -60,261 +281,7 @@ pub fn build_network_service(
         },
         ..NetworkConfiguration::default()
     };
-    NetworkService::new(config)
+    let mut service = SNetworkService::new(config, handle);
+    let (net_tx, net_rx, event_rx, control_tx) = service.run();
+    (service, net_tx, net_rx, event_rx, control_tx)
 }
-
-fn build_libp2p_service(cfg: NetworkConfiguration) -> Result<Arc<Mutex<Libp2pService>>, io::Error> {
-    let protocol = network_libp2p::ProtocolId::from("stargate".as_bytes());
-    match start_service(protocol, cfg) {
-        Ok((srv, _)) => Ok(Arc::new(Mutex::new(srv))),
-        Err(err) => Err(err.into()),
-    }
-}
-
-fn run_network(
-    net_srv: Arc<Mutex<Libp2pService>>,
-    acks: Arc<Mutex<HashMap<u128, Sender<()>>>>,
-) -> (
-    mpsc::UnboundedSender<NetworkMessage>,
-    mpsc::UnboundedReceiver<NetworkMessage>,
-    mpsc::UnboundedReceiver<PeerEvent>,
-    impl Future<Item = (), Error = std::io::Error>,
-) {
-    let (mut _tx, net_rx) = mpsc::unbounded();
-    let (net_tx, mut _rx) = mpsc::unbounded::<NetworkMessage>();
-    let (event_tx, mut event_rx) = mpsc::unbounded::<PeerEvent>();
-
-    let net_srv_2 = net_srv.clone();
-    let ack_sender = net_srv.clone();
-    let task_notify = Arc::new(AtomicTask::new());
-    let notify = task_notify.clone();
-    let network_fut = stream::poll_fn(move || {
-        notify.register();
-        net_srv_2.lock().poll()
-    })
-    .for_each(move |event| {
-        match event {
-            ServiceEvent::CustomMessage { peer_id, message } => {
-                //todo: Error handle
-                let message = Message::from_bytes(message.as_ref()).unwrap();
-                match message {
-                    Message::Payload(payload) => {
-                        //receive message
-                        info!("Receive message with peer_id:{:?}", &peer_id);
-                        let address = convert_peer_id_to_account_address(&peer_id).unwrap();
-                        let user_msg = NetworkMessage {
-                            peer_id: address,
-                            data: payload.data,
-                        };
-                        let _ = _tx.unbounded_send(user_msg);
-                        if payload.id != 0 {
-                            ack_sender.lock().send_custom_message(
-                                &peer_id,
-                                Message::ACK(payload.id).into_bytes(),
-                            );
-                        }
-                    }
-                    Message::ACK(message_id) => {
-                        info!("Receive message ack");
-                        if let Some(tx) = acks.lock().remove(&message_id) {
-                            let _ = tx.send(());
-                        } else {
-                            error!(
-                                "Receive a invalid ack, message id:{}, peer id:{}",
-                                message_id, peer_id
-                            );
-                        }
-                    }
-                }
-            }
-            ServiceEvent::OpenedCustomProtocol {
-                peer_id,
-                version: _,
-                debug_info: _,
-            } => {
-                let addr = convert_peer_id_to_account_address(&peer_id).unwrap();
-                info!("Connected peer {:?}", addr);
-                let open_msg = PeerEvent::Open(addr);
-                let _ = event_tx.unbounded_send(open_msg);
-            }
-            ServiceEvent::ClosedCustomProtocol {
-                peer_id,
-                debug_info: _,
-            } => {
-                let addr = convert_peer_id_to_account_address(&peer_id).unwrap();
-                info!("Close peer {:?}", addr);
-                let open_msg = PeerEvent::Close(addr);
-                let _ = event_tx.unbounded_send(open_msg);
-            }
-            ServiceEvent::Clogged {
-                peer_id: _,
-                messages: _,
-            } => debug!("Network clogged"),
-        };
-        Ok(())
-    })
-    .then(|_| {
-        debug!("Finish network poll");
-        Ok(())
-    });
-
-    let protocol_fut = stream::poll_fn(move || _rx.poll())
-        .for_each(move |message| {
-            let peer_id = convert_account_address_to_peer_id(message.peer_id).unwrap();
-            net_srv
-                .lock()
-                .send_custom_message(&peer_id, Message::new_message(message.data).into_bytes());
-            task_notify.notify();
-            if net_srv.lock().is_open(&peer_id) == false {
-                error!(
-                    "Message send to peer :{} is not connected",
-                    convert_peer_id_to_account_address(&peer_id).unwrap()
-                );
-            }
-            info!("Already send message {:?}", &peer_id);
-            Ok(())
-        })
-        .then(|res| {
-            match res {
-                Ok(()) => {
-                    debug!("Finish prototol poll");
-                }
-                Err(_) => error!("protocol disconnected"),
-            };
-            Ok(())
-        });
-    let futures: Vec<Box<dyn Future<Item = (), Error = io::Error> + Send>> = vec![
-        Box::new(network_fut) as Box<_>,
-        Box::new(protocol_fut) as Box<_>,
-    ];
-
-    let futs = futures::select_all(futures)
-        .and_then(move |_| {
-            debug!("Networking ended");
-            Ok(())
-        })
-        .map_err(|(r, _, _)| r);
-
-    //let futs = connected_fut.and_then(move |_| futs);
-
-    (net_tx, net_rx, event_rx, futs)
-}
-
-fn spawn_network(
-    libp2p_service: Arc<Mutex<Libp2pService>>,
-    acks: Arc<Mutex<HashMap<u128, Sender<()>>>>,
-    close_rx: oneshot::Receiver<()>,
-) -> (
-    mpsc::UnboundedSender<NetworkMessage>,
-    mpsc::UnboundedReceiver<NetworkMessage>,
-    mpsc::UnboundedReceiver<PeerEvent>,
-) {
-    let (network_sender, network_receiver, event_rx, network_future) =
-        run_network(libp2p_service, acks);
-    let fut = network_future
-        .select(close_rx.then(|_| {
-            debug!("Shutdown the network");
-            Ok(())
-        }))
-        .map(|(val, _)| val)
-        .map_err(|(_err, _)| ());
-    let mut runtime = tokio::runtime::Builder::new()
-        .name_prefix("libp2p-")
-        .build()
-        .unwrap();
-    let _thread = thread::Builder::new()
-        .name("network".to_string())
-        .spawn(move || {
-            let _ = runtime.block_on(fut);
-        });
-    (network_sender, network_receiver, event_rx)
-}
-
-impl NetworkService {
-    fn new(
-        cfg: NetworkConfiguration,
-    ) -> (
-        NetworkService,
-        mpsc::UnboundedSender<NetworkMessage>,
-        mpsc::UnboundedReceiver<NetworkMessage>,
-        mpsc::UnboundedReceiver<PeerEvent>,
-        oneshot::Sender<()>,
-    ) {
-        let (close_tx, close_rx) = oneshot::channel::<()>();
-        let libp2p_service = build_libp2p_service(cfg).unwrap();
-        let acks = Arc::new(Mutex::new(HashMap::new()));
-        let (network_sender, network_receiver, event_rx) =
-            spawn_network(libp2p_service.clone(), acks.clone(), close_rx);
-        info!("Network started, connected peers:");
-        for p in libp2p_service.lock().connected_peers() {
-            info!("peer_id:{}", p);
-        }
-
-        (
-            Self {
-                libp2p_service,
-                acks,
-            },
-            network_sender,
-            network_receiver,
-            event_rx,
-            close_tx,
-        )
-    }
-
-    pub fn is_connected(&self, address: AccountAddress) -> bool {
-        self.libp2p_service
-            .lock()
-            .is_open(&convert_account_address_to_peer_id(address).unwrap())
-    }
-
-    pub fn identify(&self) -> AccountAddress {
-        convert_peer_id_to_account_address(self.libp2p_service.lock().peer_id()).unwrap()
-    }
-
-    pub fn send_message(
-        &mut self,
-        account_address: AccountAddress,
-        message: Vec<u8>,
-    ) -> impl Future<Item = (), Error = Canceled> {
-        let (tx, rx) = oneshot::channel::<()>();
-        let (protocol_msg, message_id) = Message::new_payload(message);
-        let peer_id =
-            convert_account_address_to_peer_id(account_address).expect("Invalid account address");
-
-        self.libp2p_service
-            .lock()
-            .send_custom_message(&peer_id, protocol_msg.into_bytes());
-        debug!("Send message with ack");
-        self.acks.lock().insert(message_id, tx);
-        rx
-    }
-
-    pub fn broadcast_message(&mut self, message: Vec<u8>) {
-        debug!("start send broadcast message");
-        let (protocol_msg, message_id) = Message::new_payload(message);
-
-        let message_bytes = protocol_msg.into_bytes();
-
-        let mut peers = HashSet::new();
-
-        for p in self.libp2p_service.lock().connected_peers() {
-            debug!("will send message to {}", p);
-            peers.insert(p.clone());
-        }
-
-        for peer_id in peers {
-            self.libp2p_service
-                .lock()
-                .send_custom_message(&peer_id, message_bytes.clone());
-        }
-        debug!("finish send broadcast message");
-    }
-}
-
-pub type NetworkComponent = (
-    NetworkService,
-    mpsc::UnboundedSender<NetworkMessage>,
-    mpsc::UnboundedReceiver<NetworkMessage>,
-    oneshot::Sender<()>,
-);
