+++ network/src/sync_messages.rs
@@ -25,6 +25,8 @@ pub enum DownloadMessage {
     BatchBodyMsg(BatchBodyMsg),
     BatchHeaderAndBodyMsg(BatchHeaderMsg, BatchBodyMsg),
     NewHeadBlock(PeerInfo, Block),
+    // just fo test
+    MinedBlock(Block),
 }
 
 impl Message for DownloadMessage {
