+++ node/src/main.rs
@@ -69,7 +69,14 @@ async fn main() {
     // let mut config = NodeConfig::default();
     // config.network.listen = format!("/ip4/127.0.0.1/tcp/{}", config::get_available_port());
     // let node_config = Arc::new(config);
-    let network = NetworkActor::launch(node_config.clone(), bus.clone(), txpool.clone(), keypair);
+    let rt = tokio::runtime::Runtime::new().unwrap();
+    let network = NetworkActor::launch(
+        node_config.clone(),
+        bus.clone(),
+        txpool.clone(),
+        keypair,
+        rt.handle().clone(),
+    );
     let chain = ChainActor::launch(
         node_config.clone(),
         startup_info,
